pyproject.toml
[project]
name = "chat-history"
version = "0.1.0"
description = ""
authors = [
    {name = "rohit0221",email = "rohit0221@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "redis (>=5.2.1,<6.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "openai (>=1.64.0,<2.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
